% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCellularComponentToGraph.R
\name{addCellularComponentToGraph}
\alias{addCellularComponentToGraph}
\title{Internal function to add the CC semantic similarity attribute to the 
results graph object}
\usage{
addCellularComponentToGraph(graph = NULL, GO.CellularComponent = NULL,
  GONamesAsLabels = T, organism = "human")
}
\arguments{
\item{graph}{An \code{\link[igraph]{igraph}} object, typically a small one 
coming from an enrichment procedure}

\item{GO.CellularComponent}{Character, GO entry to draw semantic similarity 
in the solution graph. If no similarity is desired, leave it as NULL}

\item{GONamesAsLabels}{Logical, should GO names be displayed as labels instead 
of GO identifiers?}

\item{organism}{Character, organism to pass to \code{\link[GOSemSim]{goSim}}}
}
\value{
An \code{\link{igraph}} object that contains an extra attribute: \code{GO.CC}
}
\description{
Function \code{addCellularComponentToGraph} takes and returns a graph object with class 
\code{\link[igraph]{igraph}} adding the attribute \code{GO.CC} for semantic similarity.
}

