% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGOToGraph.R
\name{addGOToGraph}
\alias{addGOToGraph}
\title{Internal function to add the GO semantic similarity attribute to the 
results graph object}
\usage{
addGOToGraph(graph = NULL, GOterm = NULL, godata.options = list(OrgDb =
  "org.Hs.eg.db", ont = "CC"), mart.options = list(biomart = "ensembl",
  dataset = "hsapiens_gene_ensembl"), GONamesAsLabels = TRUE)
}
\arguments{
\item{graph}{An \code{\link[igraph]{igraph}} object, typically a small one 
coming from an enrichment procedure}

\item{GOterm}{Character, GO entry to draw 
semantic similarity in the solution graph.}

\item{godata.options}{List, options for the database creator 
\code{\link[GOSemSim]{godata}}}

\item{mart.options}{List, options for the \code{biomaRt} function
\code{\link[biomaRt]{getBM}}. Importantly, this defines the organism, 
see \code{\link[biomaRt]{listDatasets}} for possibilities}

\item{GONamesAsLabels}{Logical, should GO names 
be displayed as labels instead of GO identifiers?}
}
\value{
An \code{\link{igraph}} object that contains 
an extra attribute: \code{GO.simil}
}
\description{
Function \code{addGOToGraph} takes and returns 
a graph object with class 
\code{\link[igraph]{igraph}} adding the attribute 
\code{GO.simil} for semantic similarity.
}
\examples{
## This function is internal
attach(environment(FELLA:::addCellularComponentToGraph))
library(igraph)
data(FELLA.sample)
data(input.sample)
## Enrich input
obj <- enrich(
compounds = input.sample, 
data = FELLA.sample)
## Generate graph
g <- generateResultsGraph(
threshold = 0.1, 
object = obj, 
data = FELLA.sample)
## Add the cellular component
g.cc <- addGOToGraph(
graph = g, 
GOterm = "GO:0005739")

## Without the CC
any(V(g)$GO.simil >= 0)
## With the CC
v.cc <- unlist(V(g.cc)$GO.simil)
sum(v.cc >= 0)
table(v.cc)

}
