% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildDataFromGraph.R
\name{buildDataFromGraph}
\alias{buildDataFromGraph}
\title{Generate RData KEGG files from the curated KEGG graph}
\usage{
buildDataFromGraph(keggdata.graph = NULL, outputDir = NULL,
  matrices = c("hypergeom", "diffusion", "pagerank"),
  normality = c("diffusion", "pagerank"), dampingFactor = 0.7,
  niter = 1000)
}
\arguments{
\item{keggdata.graph}{An \code{\link[igraph]{igraph}} object generated by the function
\code{\link{buildGraphFromKEGG}}}

\item{outputDir}{Character containing the directory to save KEGG files}

\item{matrices}{A character vector, containing any of these:
\code{"hypergeom"}, \code{"diffusion"}, \code{"pagerank"}}

\item{normality}{A character vector, containing any of these:
\code{"diffusion"}, \code{"pagerank"}}

\item{dampingFactor}{Numeric value, damping factor \code{d} for PageRank}

\item{niter}{Numeric value, number of iterations to estimate the p-values
for the CC size.}
}
\value{
Return value is \code{invisible()}, but as a permanent action the
directory \code{outdir} is created and the RData KEGG files are saved.
}
\description{
Function \code{buildDataFromGraph} writes the KEGG data in a permanent directory.
This data is ready to use through the \code{\link{FELLA}} package. The
KEGG graph must provide from the function \code{\link{buildGraphFromKEGG}}.
}

