% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildGraphFromKEGG.R
\name{buildGraphFromKEGG}
\alias{buildGraphFromKEGG}
\title{Generate the KEGG graph from a local KEGG download}
\usage{
buildGraphFromKEGG(KEGGdirectory = NULL, saveLists = NULL,
  loadLists = NULL, GOterms = NULL)
}
\arguments{
\item{KEGGdirectory}{A character containing the local directory for the KEGG
downloaded resource, which must be available at the moment of execution}

\item{saveLists}{An optional character containing the local directory to save
intermediate data in a list format, which have already been parsed.
Afterwards, the graph can be rebuilt significantly faster through the lists}

\item{loadLists}{An optional character containing the local directory to rebuild the
KEGG graph from parsed lists, generated by this function}

\item{GOterms}{Logical value, should GO terms be added to the KEGG enzyme
category? This is not performed by default, but it will allow the GO
similarity analysis for the enrichment output}
}
\value{
Curated KEGG graph (class \code{\link[igraph]{igraph}})
}
\description{
Function \code{buildGraphFromKEGG} returns the curated KEGG graph, necessary
to build the other KEGG RData files and, ultimately, build the
\code{\link{FELLA.USER}} object. This procedure should only be used once, and
might take some amount of time.
}

