% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGraphFromKEGGREST.R
\name{buildGraphFromKEGGREST}
\alias{buildGraphFromKEGGREST}
\title{Generate the KEGG graph}
\usage{
buildGraphFromKEGGREST(organism = "hsa", filter.path = NULL)
}
\arguments{
\item{organism}{Character, KEGG code for the organism of interest}

\item{filter.path}{Character vector, pathways to filter. 
This is a pattern 
matched using regexp. E.g: '01100' to filter 
metabolic pathways in any species}
}
\value{
Curated KEGG graph (class \code{\link[igraph]{igraph}})
}
\description{
Function \code{buildGraphFromKEGG} returns the 
curated KEGG graph, necessary 
to build the other KEGG RData files and, ultimately, build the 
\code{\link[FELLA]{FELLA.USER}} object. 
This procedure should only be used once, and 
might take some amount of time.
}
\examples{
## These examples take several seconds to compute and require an 
## active internet conenction to connect to the KEGG API

\dontrun{
## Graph for Homo sapiens using all the pathways
library(igraph)
g <- buildGraphFromKEGGREST()
summary(g)}

\dontrun{
## Graph for Mus musculus discarding the mmu01100 pathway
g <- buildGraphFromKEGGREST(
organism = "mmu", 
filter.path = "mmu01100")}

}
