% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{D.hypergeom-class}
\alias{D.hypergeom-class}
\title{An internal S4 class for the binary matrix (hypergeometric test)}
\description{
An internal S4 class for the binary matrix (hypergeometric test)
}
\section{Slots}{

\describe{
\item{\code{matrix}}{Binary sparse matrix}
}}

