% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{D.keggdata-class}
\alias{D.keggdata-class}
\title{An internal S4 class to represent the KEGG graph and related files}
\description{
An internal S4 class to represent the KEGG graph and related files
}
\section{Slots}{

\describe{
\item{\code{graph}}{KEGG graph}

\item{\code{id2name}}{Mapping list: KEGG ID to KEGG name (can contain multiple hits)}

\item{\code{pvalues.size}}{Numeric matrix for the evaluation of CC through their size}

\item{\code{id}}{List with character vectors for KEGG categories}

\item{\code{status}}{Character that specifies the current status of this S4 class}
}}

