% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{D.pagerank-class}
\alias{D.pagerank-class}
\title{An internal S4 class for the PageRank data}
\description{
An internal S4 class for the PageRank data
}
\section{Slots}{

\describe{
\item{\code{matrix}}{Numeric (dense) matrix [optional]}

\item{\code{rowSums}}{Numeric named vector with rowSums internal data}

\item{\code{squaredRowSums}}{Numeric named vector with
squaredRowSums internal data}
}}

