% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrich}
\alias{enrich}
\title{Wrapper function for enrichment analysis}
\usage{
enrich(compounds = NULL, compoundsBackground = NULL, method = "all",
  loadMatrix = "none", approx = "normality", t.df = 10, niter = 1000,
  path = "", data = NULL, ...)
}
\arguments{
\item{compounds}{Character vector containing the KEGG IDs of the compounds
considered as affected}

\item{compoundsBackground}{Character vector containing the KEGG IDs of
the compounds that belong to the background. Can be \code{NULL} for the
default background (all compounds)}

\item{method}{Character, one of: "hypergeom", "diffusion", "pagerank" or "all"}

\item{loadMatrix}{Character to choose if heavy matrices should be loaded.
One of: "diffusion", "pagerank", "all" or NULL}

\item{approx}{Character: "simulation" for Monte Carlo, "normality",
"gamma" or "t" for parametric approaches}

\item{t.df}{Numeric value; number of degrees of freedom of the t distribution
if the approximation \code{approx = "t"} is used}

\item{niter}{Number of iterations (permutations) for Monte Carlo ("simulation"),
must be a numeric value between 1e2 and 1e5}

\item{path}{Character, path to load the \code{\link{FELLA.DATA}} object if
it is not already passed through the argument \code{data}}

\item{data}{FELLA.DATA object}
}
\value{
The \code{\link{FELLA.USER}} and the \code{\link{FELLA.DATA}} objects
in a list.
}
\description{
The function \code{\link{enrich}} is a big wrapper to perform the enrichment
analysis. It loads affected compounds, KEGG data and return a list containing
the \code{\link{FELLA.DATA}} context data and the \code{\link{FELLA.USER}}
object with the enrichment results available.
}

