% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportResults.R
\name{exportResults}
\alias{exportResults}
\title{Export the enrichment results in several file formats}
\usage{
exportResults(format = "csv", file = "myOutput", method = "diffusion",
    object = NULL, data = NULL, ...)
}
\arguments{
\item{format}{Character, one of: \code{"csv"}, \code{"igraph"}, 
\code{"png"}, [\code{"pdf"} soon]}

\item{file}{Character specifying the output 
directory for the exported files}

\item{method}{Character, exactly one of: 
\code{"hypergeom"}, \code{"diffusion"}, \code{"pagerank"}}

\item{object}{FELLA.USER object}

\item{data}{FELLA.DATA object}

\item{...}{Developer parameter, currently ignored}
}
\value{
Return value is \code{invisible()}, 
but as a permanent action the 
specified \code{file} is created.
}
\description{
Function \code{exportResults} writes the enrichment results in a 
proper layout and filetype. 
It can be a plot image (png), a table (csv), 
an \code{\link[igraph]{igraph}} object 
or a pdf report [under construction]
}
\examples{
data(FELLA.sample)
data(input.sample)
obj <- enrich(
compounds = input.sample, 
data = FELLA.sample)
out.file <- tempfile()
exportResults(
format = "csv", 
threshold = 0.1, 
file = out.file, 
object = obj, 
data = FELLA.sample)
df <- read.csv(out.file)
head(df)

}
