% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllMethods.R
\docType{class}
\name{FELLA.DATA-class}
\alias{FELLA.DATA-class}
\alias{show,FELLA.DATA-method}
\title{An S4 class to represent all the necessary KEGG data}
\usage{
\S4method{show}{FELLA.DATA}(object)
}
\arguments{
\item{object}{A \code{\link[FELLA]{FELLA.DATA}} object}
}
\value{
\code{show} returns \code{invisible()}
}
\description{
An S4 class to represent all the necessary KEGG data

\code{"show"} is an S4 method to show a FELLA.DATA object
}
\section{Slots}{

\describe{
\item{\code{keggdata}}{A \code{D.keggdata} S4 object}

\item{\code{hypergeom}}{A \code{D.hypergeom} S4 object}

\item{\code{diffusion}}{A \code{D.diffusion} S4 object}

\item{\code{pagerank}}{A \code{D.pagerank} S4 object}
}}

