% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doc-package.R
\docType{package}
\name{FELLA}
\alias{FELLA}
\alias{FELLA-package}
\title{The FELLA package}
\usage{
browseVignettes("FELLA")
?FELLA.sample
?input.sample
?enrich
}
\description{
\code{FELLA} is a metabolomics data enrichment tool that contextualises
the experimental results using KEGG reactions, enzymes, modules and pathways.
The input is introduced as a list of affected metabolites (KEGG IDs),
while the output contains a comprehensive biological network layout
that relates to the affected metabolites.
}

