% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{FELLA.sample}
\alias{FELLA.sample}
\title{FELLA.DATA sample data}
\format{a \code{FELLA.DATA} object}
\source{
Generated from a mid-2014 KEGG release
}
\usage{
\code{data(FELLA.sample)}
}
\description{
This \code{FELLA.DATA} object is a small KEGG graph object.
Despite being a small database that only contains the two metabolic pathways
hsa00010 - Glycolysis / Gluconeogenesis, and hsa00640 - Propanoate metabolism,
it is useful to play around with \code{FELLA}'s
functions. It is also used for internal testing of this package.
}

