% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateEnzymesTable.R
\name{generateEnzymesTable}
\alias{generateEnzymesTable}
\title{Generate enzyme tables with genes and GO annotations}
\usage{
generateEnzymesTable(method = "diffusion", threshold = 0.005,
  nlimit = 250, LabelLengthAtPlot = 45, capPvalues = 1e-10,
  object = NULL, data = NULL)
}
\arguments{
\item{method}{Character, either 'diffusion' or 'pagerank'}

\item{threshold}{Numeric value between 0 and 1. 
Applied when filtering KEGG nodes}

\item{nlimit}{Node limit, must be a numeric value between 1 and 1000. 
This limits the order of the solution graph}

\item{LabelLengthAtPlot}{Numeric value between 10 and 50. 
Maximum length that a label can reach when plotting the graph. 
The remaining characters will be truncated using "..."}

\item{capPvalues}{Numeric value, minimum p-value 
admitted for the readable 
formatting. Smaller p-values will be displayed 
as \code{< capPvalues} in the 
table.}

\item{object}{FELLA.USER object}

\item{data}{FELLA.DATA object}
}
\value{
A table that contains the enzymes 
along with genes and GO labels
}
\description{
Function \code{generateEnzymesTable} returns a table 
that contains the selected enzymes (with additional data) 
of a \code{\link[FELLA]{FELLA.USER}} object 
with a successful enrichment analysis.
}
\examples{
data(FELLA.sample)
data(input.sample)
obj <- enrich(
compounds = input.sample, 
data = FELLA.sample)
tab <- generateEnzymesTable(
threshold = 0.1, 
object = obj, 
data = FELLA.sample)
head(tab)

}
