% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateResultsTable.R
\name{generateResultsTable}
\alias{generateResultsTable}
\title{Generate readable tables from the enrichment results}
\usage{
generateResultsTable(method = "diffusion", threshold = 0.005, plimit = 15,
  nlimit = 250, LabelLengthAtPlot = 45, capPscores = 1e-10,
  object = NULL, data = NULL)
}
\arguments{
\item{method}{Character vector, containing some of: 
\code{"hypergeom"}, \code{"diffusion"}, \code{"pagerank"}}

\item{threshold}{Numeric value between 0 and 1. 
Applied when filtering KEGG nodes}

\item{plimit}{Pathway limit, must be a numeric value between 1 and 50}

\item{nlimit}{Node limit, must be a numeric value between 1 and 1000. 
This limits the order of the solution graph}

\item{LabelLengthAtPlot}{Numeric value between 10 and 50. 
Maximum length that a label can reach when plotting the graph. 
The remaining characters will be truncated using "..."}

\item{capPscores}{Numeric value, minimum p-score 
admitted for the readable 
formatting. Smaller p-scores will be displayed 
as \code{< capPscores}}

\item{object}{FELLA.USER object}

\item{data}{FELLA.DATA object}
}
\value{
A table that contains the KEGG graph nodes with their p-score
}
\description{
Function \code{generateResultsTable} returns a table 
that contains the results 
of a \code{\link[FELLA]{FELLA.USER}} object 
with a successful enrichment analysis.
}
\examples{
data(FELLA.sample)
data(input.sample)
obj <- enrich(
compounds = input.sample, 
data = FELLA.sample)
tab <- generateResultsTable(
threshold = 0.1, 
object = obj, 
data = FELLA.sample)
head(tab)

}
