% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBackground.R
\name{getBackground}
\alias{getBackground}
\title{Get compounds in the defined background}
\usage{
getBackground(object)
}
\arguments{
\item{object}{FELLA.USER object}
}
\value{
Vector of compounds in the background. If this vector is empty, all 
the compounds are used as background by default.
}
\description{
Extractor function for the compounds defined as background
}

