% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-.R
\name{getGraph}
\alias{getGraph}
\title{Get KEGG graph}
\usage{
getGraph(data)
}
\arguments{
\item{data}{FELLA.DATA object}
}
\value{
KEGG graph as an \pkg{igraph} object
}
\description{
Extractor function for the KEGG graph from the FELLA.DATA object
}
\examples{
data(FELLA.sample)
g <- getGraph(FELLA.sample)
class(g)
}
