% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-.R
\name{getInfo}
\alias{getInfo}
\title{Get KEGG version info}
\usage{
getInfo(data)
}
\arguments{
\item{data}{FELLA.DATA object}
}
\value{
Character containing the KEGG release details
}
\description{
Extractor function for the info about the KEGG version used to 
build the FELLA.DATA object
}
\examples{
data(FELLA.sample)
getInfo(FELLA.sample)

}
