% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getMatrix.R
\name{getMatrix}
\alias{getMatrix}
\title{Get matrix for the desired methodology}
\usage{
getMatrix(data, type)
}
\arguments{
\item{data}{FELLA.DATA object}

\item{type}{Character, one of "hypergeom", "diffusion" or "pagerank"}
}
\value{
Matrix for the desired methodology (internal usage)
}
\description{
Extractor function for the matrices of hypergeometric, diffusion and PageRank methodologies
}

