% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getName.R
\name{getName}
\alias{getName}
\title{Map KEGG identifiers to KEGG names}
\usage{
getName(data, id)
}
\arguments{
\item{data}{FELLA.DATA object}

\item{id}{KEGG IDs whose name is desired}
}
\value{
List whose names are KEGG IDs and whose entries are the vectors of matches
}
\description{
Map KEGG identifiers to KEGG names, multiple names for an ID are reported if annotated. 
The KEGG identifiers may have mixed levels.
}

