% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPscores.R
\name{getPscores}
\alias{getPscores}
\title{Get p-scores from the desired methodology}
\usage{
getPscores(object, type)
}
\arguments{
\item{object}{FELLA.USER object}

\item{type}{Character, one of "hypergeom", "diffusion" or "pagerank"}
}
\value{
Named vector of p-scores
}
\description{
Extractor function for the p-scores using the desired methodology
}
\examples{
data(FELLA.sample)
data(input.sample)
obj <- enrich(
compounds = input.sample, 
data = FELLA.sample)
p <- getPscores(obj, "diffusion")
sum(p < 0.1)

}
