% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalues.R
\name{getPvalues}
\alias{getPvalues}
\title{Get p-values from the desired methodology}
\usage{
getPvalues(object, type)
}
\arguments{
\item{object}{FELLA.USER object}

\item{type}{Character, one of "hypergeom", "diffusion" or "pagerank"}
}
\value{
Named vector of p-values
}
\description{
Extractor function for the p-values using the desired methodology
}

