% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-.R
\name{getPvaluesSize}
\alias{getPvaluesSize}
\title{Get matrix for the p-value regarding CC size}
\usage{
getPvaluesSize(data)
}
\arguments{
\item{data}{FELLA.DATA object}
}
\value{
Matrix with p-values for CC size (internal usage)
}
\description{
Extractor function for the matrix containing p-value by CC size
that compares to a random selection of nodes in the KEGG graph
}
\examples{
## This function is internal
attach(environment(FELLA:::getPvaluesSize))
data(FELLA.sample)
M <- getPvaluesSize(FELLA.sample)
dim(M)
summary(as.vector(M))
}
\keyword{internal}
