% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-.R
\name{getStatus}
\alias{getStatus}
\title{Get the slot "status"}
\usage{
getStatus(data)
}
\arguments{
\item{data}{FELLA.DATA object}
}
\value{
Slot "status" (internal usage)
}
\description{
Extractor function for the slot "status" for the KEGG data
}
\examples{
## This function is internal

data(FELLA.sample)

## Is the object loade?
FELLA:::getStatus(FELLA.sample)
FELLA:::getStatus(new("FELLA.DATA"))
}
