% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsed.R
\name{getUsed}
\alias{getUsed}
\title{Get the slot "used"}
\usage{
getUsed(object, type)
}
\arguments{
\item{object}{FELLA.USER object}

\item{type}{Character, one of "hypergeom", "diffusion" or "pagerank"}
}
\value{
Slot "used" (internal usage)
}
\description{
Extractor function for the slot "used"
}

