% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-.R
\name{getValid}
\alias{getValid}
\title{Get the slot "valid"}
\usage{
getValid(object, method)
}
\arguments{
\item{object}{FELLA.USER object}

\item{method}{Character, exactly one of: 
\code{"hypergeom"}, \code{"diffusion"}, \code{"pagerank"}}
}
\value{
Slot "valid" (internal usage)
}
\description{
Extractor function for the slot "valid"
}
\examples{
## This function is internal

data(FELLA.sample)
data(input.sample)

obj <- enrich(
compounds = input.sample, 
method = "diffusion", 
data = FELLA.sample)

## If the analysis is valid
FELLA:::getValid(obj, "diffusion")

## Otherwise
FELLA:::getValid(new("FELLA.USER"), "diffusion")
FELLA:::getValid(obj, "pagerank")

}
