% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-.R
\name{is.FELLA.DATA}
\alias{is.FELLA.DATA}
\title{Check FELLA.DATA class}
\usage{
is.FELLA.DATA(x = NULL)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical value stating if \code{x} 
is a \code{\link[FELLA]{FELLA.DATA}} object
}
\description{
Is \code{x} a \code{\link[FELLA]{FELLA.DATA}} object?
}
\examples{
data(FELLA.sample)
is.FELLA.DATA(FELLA.sample)
is.FELLA.DATA(42)
}
