% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.FELLA.USER.R
\name{is.FELLA.USER}
\alias{is.FELLA.USER}
\title{Check FELLA.USER class}
\usage{
is.FELLA.USER(x = NULL)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical value stating if \code{x} is a \code{\link{FELLA.USER}} object
}
\description{
Is \code{x} a \code{\link{FELLA.USER}} object?
}

