% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGraphFromKEGGREST.R
\name{largestcc}
\alias{largestcc}
\title{Extract largest CC}
\usage{
largestcc(graph)
}
\arguments{
\item{graph}{Igraph object}
}
\value{
Connected igraph object
}
\description{
Function \code{largestcc} extracts the 
largest connected component of an igraph object
}
\examples{
library(igraph)
g <- barabasi.game(10) + graph.empty(10)
FELLA:::largestcc(g)

}
\keyword{internal}
