% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-.R
\name{listCategories}
\alias{listCategories}
\title{List of node categories}
\usage{
listCategories()
}
\value{
Character vector
}
\description{
Node categories used in the internal representations
}
\examples{
listCategories()

}
