% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listInternalDatabases.R
\name{listInternalDatabases}
\alias{listInternalDatabases}
\title{List internal databases}
\usage{
listInternalDatabases(full.names = FALSE)
}
\arguments{
\item{full.names}{Logical, should full paths be returned?}
}
\value{
Vector with database directories
}
\description{
This function lists the directories in the local database path
}
\examples{
listInternalDatabases()

}
