\name{loadKEGGdata}
\alias{loadKEGGdata}
\title{Load KEGG data files to perform the enrichment}
\description{
Function \code{loadKEGGdata} returns a FELLA.DATA object that contains all 
the necessary biological context to perform the enrichment.
}
\usage{

loadKEGGdata(path = "data/", loadMatrix = NULL)

}

\arguments{
  \item{path}{
  Character, the path to the directory containing the RData files generated by 
  \link{buildGraphFromKEGG}.
  }
  \item{loadMatrix}{
  Character specifying whether matrices should be loaded for a faster simulated 
  p-value calculation. These matrices can consume up to 1.4GB on memory each.
  Options are NULL, "diffusion", "pagerank" and "all".
  }
}

\value{
A FELLA.DATA object, necessary for the enrichment analysis.
}

\seealso{
  \code{\link{buildGraphFromKEGG}} \code{\link{defineCompounds}}}

\examples{
KEGG.DATA <- loadKEGGdata(path = "data/", loadMatrix = "diffusion")
}


\author{Sergi Picart,
 \email{sergi.picart@upc.edu}}
