% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGraph.R
\name{mytriangle}
\alias{mytriangle}
\title{Add triangular shape to igraph plot}
\usage{
mytriangle(coords, v = NULL, params)
}
\arguments{
\item{coords, v, params}{clipping arguments, see 
\code{\link[igraph]{shapes}}}
}
\value{
Plot symbols
}
\description{
This function enables the usage of triangles as shape in 
the function \code{\link[igraph]{plot.igraph}}.
}
\examples{
## This function is internal
library(igraph)

add.vertex.shape(
"triangle", clip = shapes("circle")$clip,
plot = FELLA:::mytriangle)

g <- barabasi.game(10)
plot(
g, vertex.shape = "triangle", 
vertex.color = rainbow(vcount(g)),
vertex.size = seq(10, 20, length = vcount(g)))

}
\keyword{internal}
