% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotGraph.R
\name{plotGraph}
\alias{plotGraph}
\title{Internal function to plot a solution graph}
\usage{
plotGraph(graph = NULL, input = NULL, layout = F, NamesAsLabels = T,
  ...)
}
\arguments{
\item{graph}{Graph result that must come from diffusion or pagerank analysis}

\item{input}{Character vector, compounds in the input to be highlighted}

\item{layout}{Logical, should the plot be returned as a layout?}

\item{NamesAsLabels}{Logical, should KEGG names be displayed as labels instead
of KEGG identifiers?}

\item{...}{Additional parameters passed to \code{\link[igraph]{plot.igraph}}}
}
\value{
If \code{layout = F} then the value returned is \code{invisible()}.
Otherwise, the layout is returned, also in an invisible fashion.
}
\description{
This function plots a solution graph, tailored for the diffusion and pagerank
analysis.
}

