% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLegend.R
\name{plotLegend}
\alias{plotLegend}
\title{Internal function to add a legend to a graph plot}
\usage{
plotLegend(GO.annot = FALSE, cex = 0.75)
}
\arguments{
\item{GO.annot}{Logical, should GO annotations be included?}

\item{cex}{Numeric value, \code{cex} parameter for the function 
\code{\link[graphics]{legend}}}
}
\value{
This function is only used for its effect, 
so it returns \code{invisible()}
}
\description{
This function adds a legend to a solution plot. 
It can include the CC similarity.
}
\examples{
## This function is internal

library(igraph)
g <- barabasi.game(20)
plot(g)
FELLA:::plotLegend()
plot(g)
FELLA:::plotLegend(GO.annot = TRUE)
}
