% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGraphFromKEGGREST.R
\name{sanitise}
\alias{sanitise}
\title{ID sanitiser function}
\usage{
sanitise(x, category, organism)
}
\arguments{
\item{x}{Character vector, IDs to sanitise}

\item{category}{Character, one of: 
\code{"pathway"}, \code{"module"}, \code{"enzyme"}, 
\code{"ncbi-gene"}, \code{"reaction"}, \code{"compound"}}
}
\value{
Character vector, sanitised \code{x}
}
\description{
Sanitise KEGG identifiers
}
\examples{
sanitise(c("path:hsa00010", "path:hsa00020"), "pathway", "hsa")

}
\keyword{internal}
