% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{U.userinput-class}
\alias{U.userinput-class}
\title{An internal S4 class for the user input data}
\description{
An internal S4 class for the user input data
}
\section{Slots}{

\describe{
\item{\code{metabolites}}{Character vector containing the affected compounds}

\item{\code{metabolitesbackground}}{Character vector containing the compounds 
for the personalised background. Optionally, can be \code{NULL} for default 
background}

\item{\code{excluded}}{Character vector containing the compounds that have been 
excluded because they cannot be mapped to KEGG graph compounds}
}}

