% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDataFromGraph.R
\name{buildDataFromGraph}
\alias{buildDataFromGraph}
\title{Generate RData KEGG files from the curated KEGG graph}
\usage{
buildDataFromGraph(keggdata.graph = NULL, databaseDir = "myDatabase",
    internalDir = TRUE, matrices = c("hypergeom", "diffusion", "pagerank"),
    normality = c("diffusion", "pagerank"), dampingFactor = 0.85,
    niter = 1000)
}
\arguments{
\item{keggdata.graph}{An \code{\link[igraph]{igraph}} 
object generated by the function 
\code{\link[FELLA]{buildGraphFromKEGGREST}}}

\item{databaseDir}{Character containing the directory to save KEGG files. 
It should contain no slashes if \code{internalDir = TRUE}}

\item{internalDir}{Logical, should the directory be internal 
in the package directory?}

\item{matrices}{A character vector, containing any of these: 
\code{"hypergeom"}, \code{"diffusion"}, \code{"pagerank"}}

\item{normality}{A character vector, containing any of these: 
\code{"diffusion"}, \code{"pagerank"}}

\item{dampingFactor}{Numeric value between 0 and 1 (none inclusive), 
damping factor \code{d} for 
PageRank (\code{\link[igraph]{page_rank}})}

\item{niter}{Numeric value, number of iterations to estimate the p-values 
for the CC size. Between 10 and 1e4.}
}
\value{
Return value is \code{invisible(TRUE)} if successful,
and as a permanent action the 
directory \code{outdir} is created and the RData KEGG files are saved.
}
\description{
Function \code{buildDataFromGraph} writes the KEGG data 
in a permanent directory. 
This data is ready to use through the 
\code{\link[FELLA]{FELLA}} package. The 
KEGG graph must provide from the function 
\code{\link[FELLA]{buildGraphFromKEGGREST}}.
}
\examples{
## Toy example
data("FELLA.sample")
## Graph to build the database
g.sample <- FELLA:::getGraph(FELLA.sample)
dir.tmp <- tempdir()
## Save it in a temporary directory
buildDataFromGraph(
keggdata.graph = g.sample, 
databaseDir = dir.tmp, 
internalDir = FALSE, 
matrices = c("hypergeom", "diffusion", "pagerank"), 
normality = c("diffusion", "pagerank"), 
dampingFactor = 0.85,
niter = 10)

\dontrun{
## Build your local database
## This builds the full version, thus can take some time
## Note that setting niter higher takes time
g <- buildGraphFromKEGGREST()
buildDataFromGraph(
keggdata.graph = g, 
outputDir = "myFirstDb", 
internalDir = TRUE, 
matrices = c("hypergeom", "diffusion", "pagerank"), 
normality = c("diffusion", "pagerank"), 
dampingFactor = 0.85,
niter = 1e3)}

}
