% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGraphFromKEGGREST.R
\name{buildGraphFromKEGGREST}
\alias{buildGraphFromKEGGREST}
\title{Generate the KEGG graph}
\usage{
buildGraphFromKEGGREST(organism = "hsa", filter.path = NULL,
  GOterms_hsa = T)
}
\arguments{
\item{organism}{Character, KEGG code for the organism of interest}

\item{filter.path}{Character vector, pathways to filter. This is a pattern 
matched using regexp. E.g: '01100' to filter metabolic pathways in any species}

\item{GOterms_hsa}{Logical value, should GO terms be added to the KEGG enzyme 
category? This is performed by default, and allows the GO 
similarity analysis for the enrichment output}
}
\value{
Curated KEGG graph (class \code{\link[igraph]{igraph}})
}
\description{
Function \code{buildGraphFromKEGG} returns the curated KEGG graph, necessary 
to build the other KEGG RData files and, ultimately, build the 
\code{\link[FELLA]{FELLA.USER}} object. This procedure should only be used once, and 
might take some amount of time.
}

