% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCompounds.R
\name{defineCompounds}
\alias{defineCompounds}
\title{Define the list of input compounds}
\usage{
defineCompounds(compounds = NULL, compoundsBackground = NULL, data = NULL)
}
\arguments{
\item{compounds}{Character vector containing the KEGG IDs of the compounds
considered as affected}

\item{compoundsBackground}{Character vector containing the KEGG IDs of 
the compounds that belong to the background. Can be \code{NULL} for the 
default background (all compounds)}

\item{data}{FELLA.DATA object}
}
\value{
The \code{\link[FELLA]{FELLA.USER}} object to perform the user analysis
}
\description{
This function maps the specficied list of compounds, usually from an 
experimental metabolomics study, in the graph contained in the
\code{\link[FELLA]{FELLA.DATA}} object. Once the compounds are mapped, the enrichment 
can be performed. There is also the option to define a personalised background.
}

