% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrich}
\alias{enrich}
\title{Wrapper function for enrichment analysis}
\usage{
enrich(compounds = NULL, compoundsBackground = NULL,
    methods = listMethods(), loadMatrix = "none", approx = "normality",
    t.df = 10, niter = 1000, databaseDir = "myDatabase",
    internalDir = TRUE, data = NULL, ...)
}
\arguments{
\item{compounds}{Character vector containing the 
KEGG IDs of the compounds considered as affected}

\item{compoundsBackground}{Character vector containing the KEGG IDs of 
the compounds that belong to the background. Can be \code{NULL} for the 
default background (all compounds)}

\item{methods}{Character vector, containing some of: 
\code{"hypergeom"}, \code{"diffusion"}, \code{"pagerank"}}

\item{loadMatrix}{Character vector to choose if 
heavy matrices should be loaded. 
Can contain: \code{"diffusion"}, \code{"pagerank"}}

\item{approx}{Character: "simulation" for Monte Carlo, "normality", 
"gamma" or "t" for parametric approaches}

\item{t.df}{Numeric value; number of degrees of freedom 
of the t distribution 
if the approximation \code{approx = "t"} is used}

\item{niter}{Number of iterations (permutations) 
for Monte Carlo ("simulation"), 
must be a numeric value between 1e2 and 1e5}

\item{databaseDir}{Character, path to load the 
\code{\link[FELLA]{FELLA.DATA}} object if
it is not already passed through the argument \code{data}}

\item{internalDir}{Logical, is the directory located 
in the package directory?}

\item{data}{FELLA.DATA object}

\item{...}{Further arguments for the enrichment function(s) 
\code{\link[FELLA]{runDiffusion}}, \code{\link[FELLA]{runPagerank}}}
}
\value{
The \code{\link[FELLA]{FELLA.USER}} 
and the \code{\link[FELLA]{FELLA.DATA}} objects 
in a list.
}
\description{
The function \code{\link[FELLA]{enrich}} is a big wrapper 
to perform the enrichment analysis. It loads affected compounds, 
KEGG data and return a list containing 
the \code{\link[FELLA]{FELLA.DATA}} context data and the 
\code{\link[FELLA]{FELLA.USER}} 
object with the enrichment results available.
}
\examples{
## Load the internal database. This one is a toy example!
## Do not use as a regular database
data(FELLA.sample)
## Load a list of compounds to enrich
data(input.sample)
## Launch wrapper
obj <- enrich(
compounds = input.sample, 
data = FELLA.sample)
obj

}
