% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportResults.R
\name{exportResults}
\alias{exportResults}
\title{Export the enrichment results in several file formats}
\usage{
exportResults(format = "csv", file = "myOutput", method = "diffusion",
  threshold = 0.05, plimit = 15, nlimit = 250, object = NULL,
  data = NULL, ...)
}
\arguments{
\item{format}{Character, one of: \code{"csv"}, \code{"igraph"},
\code{"png"}, [\code{"pdf"} soon]}

\item{file}{Character specifying the output directory for the exported files}

\item{method}{Character, one of: "hypergeom", "diffusion", "pagerank" or "all"}

\item{threshold}{Numeric value between 0 and 1. Applied when filtering KEGG nodes}

\item{plimit}{Pathway limit, must be a numeric value between 1 and 50}

\item{nlimit}{Node limit, must be a numeric value between 1 and 1000. This limits
the order of the solution graph}

\item{object}{FELLA.USER object}

\item{data}{FELLA.DATA object}
}
\value{
Return value is \code{invisible()}, but as a permanent action the
specified \code{file} is created.
}
\description{
Function \code{exportResults} writes the enrichment results in a
proper layout and filetype. It can be a plot image (png), a table (csv),
an \code{\link[igraph]{igraph}} object or a pdf report [under construction]
}

