% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-package.R
\docType{package}
\name{FELLA}
\alias{FELLA}
\alias{FELLA-package}
\title{The FELLA package}
\description{
\code{FELLA} is a metabolomics data enrichment tool that contextualises 
the experimental results using KEGG reactions, 
enzymes, modules and pathways. 
The input is introduced as a list of affected metabolites (KEGG IDs), 
while the output contains a comprehensive biological network layout 
that relates to the affected metabolites.
}
\examples{
browseVignettes("FELLA")
?FELLA.sample
?input.sample
?enrich

}
