% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllMethods.R
\docType{class}
\name{FELLA.USER-class}
\alias{FELLA.USER-class}
\alias{plot,FELLA.USER,ANY-method}
\alias{show,FELLA.USER-method}
\alias{summary,FELLA.USER-method}
\title{An S4 class to save all the user analysis data}
\usage{
\S4method{summary}{FELLA.USER}(object)

\S4method{show}{FELLA.USER}(object)

\S4method{plot}{FELLA.USER,ANY}(x = 1, method = "hypergeom",
  main = "Affected subgraph", threshold = 0.005, plimit = 15,
  nlimit = 250, layout = F, filename = NULL,
  splitByConnectedComponent = F, askPlots = T,
  thresholdConnectedComponent = 0.05, GO.CellularComponent = NULL,
  GONamesAsLabels = T, LabelLengthAtPlot = 22, data = NULL, ...)
}
\arguments{
\item{object}{A \code{\link{FELLA.USER}} object}

\item{x}{A \code{\link{FELLA.USER}} object}

\item{method}{Character, one of: "hypergeom", "diffusion" or "pagerank"}

\item{main}{Character; plot title}

\item{threshold}{Numeric value between 0 and 1. Applied when filtering KEGG nodes}

\item{plimit}{Pathway limit, must be a numeric value between 1 and 50}

\item{nlimit}{Node limit, must be a numeric value between 1 and 1000. This limits
the order of the solution graph}

\item{layout}{Logical, should the plot be returned as a layout?}

\item{filename}{Character; optional file name to save the plot}

\item{splitByConnectedComponent}{Logical, should the solution be split by CC?}

\item{askPlots}{Logical, should R ask for the next plots using hit<ENTER>?}

\item{GO.CellularComponent}{Character, GO entry to draw semantic similarity
in the solution graph. If no similarity is desired, leave it as NULL}

\item{GONamesAsLabels}{Logical, should GO names be displayed as labels instead
of GO identifiers?}

\item{LabelLengthAtPlot}{Numeric value between 10 and 50. Maximum length that
a label can reach when plotting the graph. The remaining characters will be
truncated using "..."}

\item{data}{FELLA.DATA object}

\item{...}{Additional arguments passed to plotting functions}

\item{object}{A \code{\link{FELLA.USER}} object}
}
\value{
\code{summary} returns a list with the summary data

\code{show} returns \code{invisible()}

\code{plot} returns a layout if \code{layout = T},
otherwise \code{invisible()}
}
\description{
An S4 class to save all the user analysis data

\code{"summary"} is an S4 method to show a summary of the FELLA.USER object

\code{"show"} is an S4 method to show a FELLA.USER object

\code{"plot"} is an S4 method to plot the results from a FELLA.USER object
}
\section{Slots}{

\describe{
\item{\code{userinput}}{A \code{U.userinput} S4 object}

\item{\code{hypergeom}}{A \code{U.hypergeom} S4 object}

\item{\code{diffusion}}{A \code{U.diffusion} S4 object}

\item{\code{pagerank}}{A \code{U.pagerank} S4 object}
}}

