% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateResultsGraph.R
\name{generateResultsGraph}
\alias{generateResultsGraph}
\title{Generate graph objects from the enrichment results}
\usage{
generateResultsGraph(method = "diffusion", threshold = 0.05, plimit = 15,
  nlimit = 250, splitByConnectedComponent = F,
  thresholdConnectedComponent = 0.05, GO.CellularComponent = NULL,
  GONamesAsLabels = T, LabelLengthAtPlot = 22, object = NULL,
  data = FELLA.DATA)
}
\arguments{
\item{method}{Character, one of: "hypergeom", "diffusion" or "pagerank"}

\item{threshold}{Numeric value between 0 and 1. Applied when filtering KEGG nodes}

\item{plimit}{Pathway limit, must be a numeric value between 1 and 50}

\item{nlimit}{Node limit, must be a numeric value between 1 and 1000. This limits
the order of the solution graph}

\item{splitByConnectedComponent}{Logical, should the solution be split by CC?}

\item{thresholdConnectedComponent}{Numeric value between 0 and 1, refers to the
threshold to keep a CC regarding its size}

\item{GO.CellularComponent}{Character, GO entry to draw semantic similarity
in the solution graph. If no similarity is desired, leave it as NULL}

\item{GONamesAsLabels}{Logical, should GO names be displayed as labels instead
of GO identifiers?}

\item{LabelLengthAtPlot}{Numeric value between 10 and 50. Maximum length that
a label can reach when plotting the graph. The remaining characters will be
truncated using "..."}

\item{object}{FELLA.USER object}

\item{data}{FELLA.DATA object}
}
\value{
An \code{\link{igraph}} object if \code{splitByConnectedComponent = F};
a list of \code{\link{igraph}} objects otherwise.
}
\description{
Function \code{generateResultsGraph} returns a graph object with class
\code{\link[igraph]{igraph}} or a list of graphs of the same class.
according to the specified threshold. A \code{\link{FELLA.USER}} object
with a successful enrichment analysis must be supplied.
}

