% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExcluded.R
\name{getExcluded}
\alias{getExcluded}
\title{Get excluded compounds}
\usage{
getExcluded(object)
}
\arguments{
\item{object}{FELLA.USER object}
}
\value{
Vector of the excluded compounds
}
\description{
Extractor function for the compounds in the input that were not mapped to the KEGG graph
}

