% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGraph.R
\name{getGraph}
\alias{getGraph}
\title{Get KEGG graph}
\usage{
getGraph(data)
}
\arguments{
\item{data}{FELLA.DATA object}
}
\value{
KEGG graph as an \link[igraph]{igraph} object
}
\description{
Extractor function for the KEGG graph from the FELLA.DATA object
}

