% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getInput.R
\name{getInput}
\alias{getInput}
\title{Get metabolites in the input}
\usage{
getInput(object)
}
\arguments{
\item{object}{FELLA.USER object}
}
\value{
Vector of metabolites in the input
}
\description{
Extractor function for the metabolites specified by the user in the input
}

