% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatrix.R
\name{getMatrix}
\alias{getMatrix}
\title{Get matrix for the desired methodology}
\usage{
getMatrix(data, type)
}
\arguments{
\item{data}{FELLA.DATA object}

\item{type}{Character, one of "hypergeom", "diffusion" or "pagerank"}
}
\value{
Matrix for the desired methodology (internal usage)
}
\description{
Extractor function for the matrices of 
hypergeometric, diffusion and PageRank methodologies
}
\examples{
## This function is internal
attach(environment(FELLA:::getMatrix))
data(FELLA.sample)
# When a matrix is loaded:
x <- getMatrix(FELLA.sample, "hypergeom")
dim(x)
# When it is not:
y <- getMatrix(FELLA.sample, "diffusion")
dim(y)
y
}
