% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSums.R
\name{getSums}
\alias{getSums}
\title{Get rowSums/squaredRowSums}
\usage{
getSums(data, type, squared)
}
\arguments{
\item{data}{FELLA.DATA object}

\item{type}{Character, one of "hypergeom", "diffusion" or "pagerank"}

\item{squared}{Logical, whether to return 
\code{rowSums} (\code{F}) or \code{squaredRowSums} (\code{T})}
}
\value{
Named vector with rowSums/squaredRowSums (internal usage)
}
\description{
Extractor function for rowSums/squaredRowSums
}
\examples{
## This function is internal
attach(environment(FELLA:::getSums))
data(FELLA.sample)
rowsums <- getSums(FELLA.sample, "diffusion", squared = FALSE)
hist(rowsums)
}

