% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValid.R
\name{getValid}
\alias{getValid}
\title{Get the slot "valid"}
\usage{
getValid(object, type)
}
\arguments{
\item{object}{FELLA.USER object}

\item{type}{Character, one of "hypergeom", "diffusion" or "pagerank"}
}
\value{
Slot "valid" (internal usage)
}
\description{
Extractor function for the slot "valid"
}
\examples{
## This function is internal
attach(environment(FELLA:::getValid))

data(FELLA.sample)
data(input.sample)

obj <- enrich(
compounds = input.sample, 
method = "diffusion", 
data = FELLA.sample)

## If the analysis is valid
getValid(obj, "diffusion")

## Otherwise
getValid(new("FELLA.USER"), "diffusion")
getValid(obj, "pagerank")

}

