% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{input.sample}
\alias{input.sample}
\title{A randomly generated list of affected metabolites}
\format{A \code{character} vector containing 30 KEGG IDs}
\source{
Generated from a mid-2014 KEGG release
}
\usage{
\code{data(input.sample)}
}
\description{
This \code{character} vector object has been generated using the
sample data in the object \code{FELLA.sample}. The KEGG compounds have been chosen
with preference for the hsa00640 pathway, so that the enrichment results choose
pathway hsa00640 over hsa00010.
}

