% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-.R
\name{is.FELLA.USER}
\alias{is.FELLA.USER}
\title{Check FELLA.USER class}
\usage{
is.FELLA.USER(x = NULL)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical value stating if \code{x} 
is a \code{\link[FELLA]{FELLA.USER}} object
}
\description{
Is \code{x} a \code{\link[FELLA]{FELLA.USER}} object?
}
\examples{
is.FELLA.USER(new("FELLA.USER"))
is.FELLA.USER(42)

data(FELLA.sample)
data(input.sample)
obj <- enrich(
compounds = input.sample, 
method = "diffusion", 
data = FELLA.sample)
is.FELLA.USER(obj)

}
