% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch a shiny app with FELLA}
\usage{
launchApp(port = 8888, host = "sp-upc", launch.browser = FALSE, ...)
}
\arguments{
\item{port, host, launch.browser}{Port and host to launch the app, 
and whether to launch browser or not}

\item{...}{Additional parameters passed to \code{\link[shiny]{runApp}}}
}
\value{
\code{invisible()}, but as a side effect the app will be launched
}
\description{
Launch a shiny application to perform the metabolomics data enrichment.
}
\examples{
r <- try(launchApp())

}

