% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadKEGGdata.R
\name{loadKEGGdata}
\alias{loadKEGGdata}
\title{Load KEGG data}
\usage{
loadKEGGdata(path = "", loadMatrix = NULL)
}
\arguments{
\item{path}{Path for the KEGG RData files}

\item{loadMatrix}{Character to choose if heavy matrices should be loaded. 
One of: "diffusion", "pagerank", "all" or NULL}
}
\value{
The \code{\link[FELLA]{FELLA.DATA}} object that contains the KEGG representation
}
\description{
This function loads all necessary contextual data from KEGG as a 
\code{\link[FELLA]{FELLA.DATA}} object. This object is necessary to perform any 
kind of enrichment using \code{\link[FELLA]{FELLA}}.
}

