% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGraph.R
\name{plotGraph}
\alias{plotGraph}
\title{Internal function to plot a solution graph}
\usage{
plotGraph(graph = NULL, layout = FALSE, graph.layout = NULL,
    plotLegend = TRUE, plot.fun = "plot.igraph", NamesAsLabels = TRUE, ...)
}
\arguments{
\item{graph}{Graph result that must come 
from diffusion or pagerank analysis}

\item{layout}{Logical, should the plot be returned as a layout?}

\item{graph.layout}{Two-column numeric matrix, if this argument is not null 
then it is used as graph layout}

\item{plotLegend}{Logical, should the legend be plotted as well?}

\item{plot.fun}{Character, can be either \code{plot.igraph} or \code{tkplot}}

\item{NamesAsLabels}{Logical, should KEGG names be displayed 
as labels instead of KEGG identifiers?}

\item{...}{Optional further arguments for the plotting function}
}
\value{
If \code{layout = F} then the value 
returned is \code{invisible()}. 
Otherwise, the layout is returned, 
also in an invisible fashion.
}
\description{
This function plots a solution graph, 
tailored for the diffusion and pagerank
analysis.
}
\examples{
## This function is internal
data(FELLA.sample)
data(input.sample)
## Enrich input
obj <- enrich(
compounds = input.sample, 
data = FELLA.sample)
## Generate graph
g <- generateResultsGraph(
threshold = 0.1, 
object = obj, 
data = FELLA.sample)
## Plot it
FELLA:::plotGraph(g)

}
