% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDiffusion.R
\name{runDiffusion}
\alias{runDiffusion}
\title{Pathway enrichment through heat diffusion}
\usage{
runDiffusion(object = NULL, data = NULL, approx = "simulation",
  t.df = 10, niter = 1000, p.adjust = "fdr", BIMODAL = F)
}
\arguments{
\item{object}{FELLA.USER object}

\item{data}{FELLA.DATA object}

\item{approx}{Character: "simulation" for Monte Carlo, "normality", 
"gamma" or "t" for parametric approaches}

\item{t.df}{Numeric value; number of degrees of freedom of the t distribution 
if the approximation \code{approx = "t"} is used}

\item{niter}{Number of iterations (permutations) for Monte Carlo ("simulation"), 
must be a numeric value between 1e2 and 1e5}

\item{p.adjust}{Character passed to the \code{\link[stats]{p.adjust}} method}

\item{BIMODAL}{Logical value, should the bimodal test for compounds be used?}
}
\value{
The \code{\link[FELLA]{FELLA.USER}} object with the diffusion enrichment results
}
\description{
Function \code{runDiffusion} performs the diffusion-based enrichment on a 
\code{\link[FELLA]{FELLA.USER}} object. If a custom background was specified, 
it will be used. 
This procedure gives statistical significance measures for each node and allows 
the extraction of a subgraph according to a fixed threshold.
}

