% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runHypergeom.R
\name{runHypergeom}
\alias{runHypergeom}
\title{Pathway enrichment through hypergeometric test}
\usage{
runHypergeom(object = NULL, data = FELLA.DATA, p.adjust = "fdr")
}
\arguments{
\item{object}{FELLA.USER object}

\item{data}{FELLA.DATA object}

\item{p.adjust}{Character passed to the \code{\link[stats]{p.adjust}} method}
}
\value{
The \code{\link{FELLA.USER}} object with the hypergeometric test results
}
\description{
Function \code{runHypergeom} performs the over representation analysis through
the hypergeometric test on a \code{\link{FELLA.USER}} object.
If a custom background was specified, it will be used.
By default, this test has smaller background than \code{\link{runDiffusion}}
and \code{\link{runPagerank}}.
}

