% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPagerank.R
\name{runPagerank}
\alias{runPagerank}
\title{Pathway enrichment through PageRank}
\usage{
runPagerank(object = NULL, data = NULL, approx = "normality",
  dampingFactor = 0.85, t.df = 10, niter = 1000, p.adjust = "fdr")
}
\arguments{
\item{object}{FELLA.USER object}

\item{data}{FELLA.DATA object}

\item{approx}{Character: "simulation" for Monte Carlo, "normality", 
"gamma" or "t" for parametric approaches}

\item{dampingFactor}{Numeric value between 0 and 1 (none inclusive), 
damping factor \code{d} for 
PageRank (\code{\link[igraph]{page_rank}})}

\item{t.df}{Numeric value; number of degrees of freedom of the t distribution 
if the approximation \code{approx = "t"} is used}

\item{niter}{Number of iterations (permutations) 
for Monte Carlo ("simulation"), 
must be a numeric value between 1e2 and 1e5}

\item{p.adjust}{Character passed to the \code{\link[stats]{p.adjust}} method}
}
\value{
The \code{\link[FELLA]{FELLA.USER}} object 
with the PageRank enrichment results
}
\description{
Function \code{runPagerank} performs the random walk 
based enrichment on a 
\code{\link[FELLA]{FELLA.USER}} object, making use of igraph function 
\code{\link[igraph]{page.rank}}. 
If a custom background was specified, it will be used. 
This procedure gives statistical significance measures for each node 
and allows the extraction of a subgraph according to a fixed threshold.
}
\examples{
data(FELLA.sample)
## Load a list of compounds to enrich
data(input.sample)
obj.empty <- defineCompounds(
compounds = input.sample, 
data = FELLA.sample)
obj.diff <- runPagerank(
object = obj.empty, 
approx = "normality", 
data = FELLA.sample)
obj.diff

## Note that the enrich wrapper can do this in a compact way
obj.diff <- enrich(
compounds = input.sample, 
method = "pagerank", 
data = FELLA.sample)
obj.diff

}

