% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showLegend.R
\name{showLegend}
\alias{showLegend}
\title{Internal function to add a legend to a graph plot}
\usage{
showLegend(GO.CellularComponent = FALSE, cex = 0.75)
}
\arguments{
\item{GO.CellularComponent}{Character, GO entry to draw 
semantic similarity in the solution graph. 
If no similarity is desired, leave it as NULL}

\item{cex}{Numeric value, \code{cex} parameter for the function 
\code{\link[graphics]{legend}}}
}
\value{
This function is only used for its effect, 
so it returns \code{invisible()}
}
\description{
This function adds a legend to a solution plot. 
It can include the CC similarity.
}
\examples{
## This function is internal
attach(environment(FELLA:::showLegend))

library(igraph)
g <- barabasi.game(20)
plot(g)
showLegend()
plot(g)
showLegend(GO.CellularComponent = TRUE)
}
